+++ backend/flaskr/__init__.py
@@ -24,15 +24,8 @@ def create_app(test_config=None):
     app = Flask(__name__)
     app.config['JSON_SORT_KEYS'] = False # we decide the order
     setup_db(app)
-
-    """
-    @TODO: Set up CORS. Allow '*' for origins. Delete the sample route after completing the TODOs
-    """
     CORS(app)
 
-    """
-    @TODO: Use the after_request decorator to set Access-Control-Allow
-    """
     @app.after_request
     def after_request(response):
         response.headers.add(
@@ -47,6 +40,7 @@ def create_app(test_config=None):
     def retrieve_categories():
         """
         Endpoint to handle GET requests for all available categories.
+        
         GET '/categories'
         {
           'categories': { 
@@ -77,6 +71,7 @@ def create_app(test_config=None):
         """
         Endpoint to handle GET requests for questions, including pagination (every 10 questions).
         This endpoint returns a list of questions, number of total questions, current category, categories.
+        
         GET '/questions?page=${integer}'
         {
           'questions': [
@@ -119,12 +114,12 @@ def create_app(test_config=None):
     @app.route('/questions/<int:question_id>', methods=["DELETE"])
     def delete_question(question_id):
         """
-        Create an endpoint to DELETE question using a question ID.
+        Endpoint to DELETE question using a question ID.
+        Deletes a specified question using the id of the question
+        Request Arguments: id - integer
+        Returns: Does not need to return anything besides the appropriate HTTP status code.
         This removal will persist in the database and when you refresh the page.
 
-        - Deletes a specified question using the id of the question
-        - Request Arguments: id - integer
-        - Returns: Does not need to return anything besides the appropriate HTTP status code. @TODO: Optionally can return the id of the question. If you are able to modify the frontend, you can have it remove the question using the id instead of refetching the questions.
         DELETE '/questions/${id}'
         """
         try:
@@ -137,37 +132,102 @@ def create_app(test_config=None):
         except:
             abort(422)
 
-    """
-    @TODO:
-    Create an endpoint to POST a new question,
-    which will require the question and answer text,
-    category, and difficulty score.
+    def create_question(question, answer, difficulty, category):
+        """
+        NB: Called from create_question_or_search_questions "shared" app route /questions - see below
 
-    TEST: When you submit a question on the "Add" tab,
-    the form will clear and the question will appear at the end of the last page
-    of the questions list in the "List" tab.
-    """
+        Endpoint to POST a new question, which will require the question and answer text, category, and difficulty score.
+        Sends a post request in order to add a new question
+        Returns: Does not return any new data
 
-    """
-    @TODO:
-    Create a POST endpoint to get questions based on a search term.
-    It should return any questions for whom the search term
-    is a substring of the question.
+        POST '/questions'
+        request body:
+        {
+            'question':  'Heres a new question string',
+            'answer':  'Heres a new answer string',
+            'difficulty': 1,
+            'category': 3,
+        }
+        """
+        try:
+            question=Question(question, answer, category, difficulty)
+            question.insert()
+            return (jsonify({"success": True}), 200)
+        except:
+            abort(422)
 
-    TEST: Search by any phrase. The questions list will update to include
-    only question that include that string within their question.
-    Try using the word "title" to start.
-    """
+    def search_questions(search_term, category):
+        """
+        NB: Called from create_question_or_search_questions "shared" app route /questions - see below
+
+        Endpoint to get questions based on a search term via a POST.
+        It returns any questions for whom the search term is a substring of the question.
+        Sends a post request in order to search for a specific question by search term
+        Returns: any array of questions, a number of totalQuestions that met the search term and the current category string
+        
+        POST '/questions'
+        request body:
+        {
+            'searchTerm': 'this is the term the user is looking for'
+        }
+        reply:
+        {
+            'questions': [
+                {
+                    'id': 1,
+                    'question': 'This is a question',
+                    'answer': 'This is an answer',
+                    'difficulty': 5,
+                    'category': 5
+                },
+                ...
+            ],
+            'totalQuestions': 1,
+            'currentCategory': 'Entertainment'
+        }
+        """
+        questions=Question.query.order_by(Question.category).order_by(Question.id).filter(Question.question.ilike('%{}%'.format(search_term))).all()
+        page_questions=paginate_questions(request, questions)
+        total_questions=len(questions)
+        current_category=category
+        return jsonify(
+            {
+                'questions': page_questions,
+                'totalQuestions': total_questions,
+                'currentCategory': current_category
+            }
+        )
+    
+    @app.route('/questions', methods=["POST"])
+    def create_question_or_search_questions():
+        """
+        Shared endpoint (app route) for create_question or search_questions - see above
+        """
+        body = request.get_json()
+        search_term = body.get("searchTerm", None)
+        if search_term is None:
+            # Create question
+            new_question = body.get("question", None)
+            new_answer = body.get("answer", None)
+            new_difficulty = body.get("difficulty", None)
+            new_category = body.get("category", None)
+            if new_question is None or new_answer is None or new_difficulty is None or new_category is None:
+                abort(422)
+            else:
+                return create_question(new_question, new_answer, new_difficulty, new_category)
+        else:
+            # Search questions
+            category = body.get("category", None)
+            return search_questions(search_term, category)
 
     @app.route('/categories/<category_id>/questions')
     def retrieve_category_questions(category_id):
         """
         Endpoint to get questions based on category.
+        Fetches questions for a cateogry specified by id request argument
+        Request Arguments: id - integer
+        Returns: An object with questions for the specified category, total questions, and current category string
 
-
-        - Fetches questions for a cateogry specified by id request argument
-        - Request Arguments: id - integer
-        - Returns: An object with questions for the specified category, total questions, and current category string
         GET '/categories/${id}/questions'
         {
           'questions': [
@@ -198,17 +258,53 @@ def create_app(test_config=None):
                 }
             )
 
-    """
-    @TODO:
-    Create a POST endpoint to get questions to play the quiz.
-    This endpoint should take category and previous question parameters
-    and return a random questions within the given category,
-    if provided, and that is not one of the previous questions.
+    @app.route('/quizzes', methods=["POST"])
+    def retrieve_next_question():
+        """
+        Endpoint to get questions to play the quiz.
+        This endpoint takes category and previous question parameters and return a random questions within the given category, if provided, and that is not one of the previous questions.
+        Sends a post request in order to get the next question
+        Returns: a single new question object
 
-    TEST: In the "Play" tab, after a user selects "All" or a category,
-    one question at a time is displayed, the user is allowed to answer
-    and shown whether they were correct or not.
-    """
+        POST '/quizzes'
+        request body:
+        {
+            'previous_questions': [1, 4, 20, 15],
+            'quiz_category': {
+                'type': category_name, 
+                'id': category_id
+            }
+        }
+        reply:
+        {
+            'question': {
+                'id': 1,
+                'question': 'This is a question',
+                'answer': 'This is an answer',
+                'difficulty': 5,
+                'category': 4
+            }
+        }
+        """
+        body = request.get_json()
+        previous_questions = body.get("previous_questions", None)
+        quiz_category = body.get("quiz_category", None)
+        if quiz_category is None:
+            abort(422)
+        else:
+            try:
+                category_id = quiz_category['id']
+                if category_id == 0: # ALL
+                    question = Question.query.filter(Question.id.not_in(previous_questions)).first()
+                else:
+                    question = Question.query.filter(Question.category==str(category_id)).filter(Question.id.not_in(previous_questions)).first()
+                return jsonify(
+                    {
+                        'question': question.format()
+                    }
+                )
+            except:
+                abort(422)
 
     """
     Error handlers for all expected errors - including 404 and 422.
@@ -235,4 +331,3 @@ def create_app(test_config=None):
         )
 
     return app
-